chore: smol touchups
