+++ crates/transaction-pool/src/pool/best.rs
@@ -2,11 +2,13 @@ use crate::{
     identifier::TransactionId, pool::pending::PendingTransaction, PoolTransaction,
     TransactionOrdering, ValidPoolTransaction,
 };
+use core::fmt;
 use reth_primitives::B256 as TxHash;
 use std::{
     collections::{BTreeMap, BTreeSet, HashSet},
     sync::Arc,
 };
+
 use tokio::sync::broadcast::{error::TryRecvError, Receiver};
 use tracing::debug;
 
@@ -187,6 +189,69 @@ impl<T: TransactionOrdering> Iterator for BestTransactions<T> {
                 return Some(best.transaction)
             }
         }
+    }
+}
+
+/// A[`BestTransactions`] implementations that filters the transactions of iter with predicate.
+///
+/// Filter out transactions are marked as invalid:
+/// [BestTransactions::mark_invalid](crate::traits::BestTransactions::mark_invalid).
+pub struct BestTransactionFilter<I, P> {
+    pub(crate) best: I,
+    pub(crate) predicate: P,
+}
+
+impl<I, P> BestTransactionFilter<I, P> {
+    /// Create a new [`BestTransactionFilter`] with the given predicate.
+    pub(crate) fn new(best: I, predicate: P) -> Self {
+        Self { best, predicate }
+    }
+}
+
+impl<I, P> Iterator for BestTransactionFilter<I, P>
+where
+    I: crate::traits::BestTransactions,
+    P: FnMut(&<I as Iterator>::Item) -> bool,
+{
+    type Item = <I as Iterator>::Item;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        loop {
+            let best = self.best.next()?;
+            if (self.predicate)(&best) {
+                return Some(best);
+            } else {
+                self.best.mark_invalid(&best);
+            }
+        }
+    }
+}
+
+impl<I, P> crate::traits::BestTransactions for BestTransactionFilter<I, P>
+where
+    I: crate::traits::BestTransactions,
+    P: FnMut(&<I as Iterator>::Item) -> bool + Send,
+{
+    fn mark_invalid(&mut self, tx: &Self::Item) {
+        crate::traits::BestTransactions::mark_invalid(&mut self.best, tx)
+    }
+
+    fn no_updates(&mut self) {
+        self.best.no_updates()
+    }
+
+    fn skip_blobs(&mut self) {
+        self.set_skip_blobs(true)
+    }
+
+    fn set_skip_blobs(&mut self, skip_blobs: bool) {
+        self.best.set_skip_blobs(skip_blobs)
+    }
+}
+
+impl<I: fmt::Debug, P> fmt::Debug for BestTransactionFilter<I, P> {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.debug_struct("BestTransactionFilter").field("best", &self.best).finish()
     }
 }
