+++ crates/transaction-pool/src/pool/mod.rs
@@ -95,9 +95,6 @@ use std::{
 use tokio::sync::mpsc;
 use tracing::{debug, trace, warn};
 mod events;
-pub use events::{FullTransactionEvent, TransactionEvent};
-
-mod listener;
 use crate::{
     blobstore::BlobStore,
     metrics::BlobStoreMetrics,
@@ -106,14 +103,16 @@ use crate::{
     validate::ValidTransaction,
 };
 use alloy_rlp::Encodable;
+pub use best::BestTransactionFilter;
+pub use blob::{blob_tx_priority, fee_delta};
+pub use events::{FullTransactionEvent, TransactionEvent};
 pub use listener::{AllTransactionsEvents, TransactionEvents};
 pub use parked::{BasefeeOrd, ParkedOrd, ParkedPool};
 pub use pending::PendingPool;
 
 mod best;
-mod best_filter;
 mod blob;
-pub use blob::{blob_tx_priority, fee_delta};
+mod listener;
 mod parked;
 pub(crate) mod pending;
 pub(crate) mod size;
