+++ crates/transaction-pool/src/traits.rs
@@ -21,7 +21,7 @@ use std::{
 };
 use tokio::sync::mpsc::Receiver;
 
-use crate::blobstore::BlobStoreError;
+use crate::{blobstore::BlobStoreError, pool::BestTransactionFilter};
 use reth_primitives::kzg::KzgSettings;
 #[cfg(feature = "serde")]
 use serde::{Deserialize, Serialize};
@@ -654,6 +654,20 @@ pub trait BestTransactions: Iterator + Send {
     ///
     /// If set to true, no blob transactions will be returned.
     fn set_skip_blobs(&mut self, skip_blobs: bool);
+
+    /// Creates an iterator which uses a closure to determine if a transaction should be yielded.
+    ///
+    /// Given an element the closure must return true or false. The returned iterator will yield
+    /// only the elements for which the closure returns true.
+    ///
+    /// Descendant transactions will be skipped.
+    fn filter<P>(self, predicate: P) -> BestTransactionFilter<Self, P>
+    where
+        P: FnMut(&Self::Item) -> bool,
+        Self: Sized,
+    {
+        BestTransactionFilter::new(self, predicate)
+    }
 }
 
 /// A no-op implementation that yields no transactions.
